feat: DetailFooter 통신로직 커스텀훅으로 분리
